fix: automated linting and formatting
