+++ src/rxiv_maker/engines/abstract.py
@@ -139,15 +139,172 @@ class AbstractContainerEngine(ABC):
         """Return the name of the container engine (e.g., 'docker', 'podman')."""
         pass
 
-    @abstractmethod
     def check_available(self) -> bool:
-        """Check if the container engine is available and running."""
-        pass
+        """Check if the container engine is available and running.
 
-    @abstractmethod
+        Returns:
+            True if engine is available and running, False otherwise.
+
+        Raises:
+            ContainerEngineNotFoundError: If engine binary is not found
+            ContainerEngineNotRunningError: If engine daemon/service is not running
+            ContainerPermissionError: If permission denied accessing engine
+            ContainerTimeoutError: If engine commands timeout
+        """
+        import logging
+        import subprocess
+
+        logger = logging.getLogger(__name__)
+
+        # Import exceptions here to avoid circular imports
+        from .exceptions import (
+            ContainerEngineNotFoundError,
+            ContainerEngineNotRunningError,
+            ContainerPermissionError,
+            ContainerTimeoutError,
+        )
+
+        try:
+            # First check if engine binary exists
+            version_result = subprocess.run([self.engine_name, "--version"], capture_output=True, text=True, timeout=5)
+
+            if version_result.returncode != 0:
+                if "permission denied" in version_result.stderr.lower():
+                    raise ContainerPermissionError(self.engine_name, f"check {self.engine_name.title()} version")
+                else:
+                    logger.debug(f"{self.engine_name.title()} version check failed: {version_result.stderr}")
+                    return False
+
+        except FileNotFoundError as e:
+            raise ContainerEngineNotFoundError(self.engine_name) from e
+        except subprocess.TimeoutExpired as e:
+            raise ContainerTimeoutError(self.engine_name, "version check", 5) from e
+
+        try:
+            # Then check if engine daemon/service is actually running
+            ps_result = subprocess.run([self.engine_name, "ps"], capture_output=True, text=True, timeout=10)
+
+            if ps_result.returncode != 0:
+                stderr_lower = ps_result.stderr.lower()
+                if "permission denied" in stderr_lower or "access denied" in stderr_lower:
+                    raise ContainerPermissionError(self.engine_name, "list containers")
+                elif "cannot connect" in stderr_lower or "connection refused" in stderr_lower:
+                    raise ContainerEngineNotRunningError(self.engine_name)
+                elif self._is_service_not_running_error(stderr_lower):
+                    raise ContainerEngineNotRunningError(self.engine_name)
+                else:
+                    logger.debug(f"{self.engine_name.title()} ps failed: {ps_result.stderr}")
+                    return False
+
+            return True
+
+        except subprocess.TimeoutExpired as e:
+            raise ContainerTimeoutError(self.engine_name, "daemon/service connectivity check", 10) from e
+        except subprocess.CalledProcessError as e:
+            logger.debug(f"{self.engine_name.title()} ps command failed with exit code {e.returncode}")
+            return False
+
+    def _is_service_not_running_error(self, stderr_lower: str) -> bool:
+        """Check if the error indicates the container service is not running.
+
+        This method can be overridden by subclasses to provide engine-specific
+        error pattern matching.
+
+        Args:
+            stderr_lower: The stderr output in lowercase
+
+        Returns:
+            True if the error indicates service is not running
+        """
+        # Common patterns for both Docker and Podman
+        service_not_running_patterns = [
+            "daemon" in stderr_lower and "not running" in stderr_lower,
+            "service" in stderr_lower and "not running" in stderr_lower,
+            "machine" in stderr_lower and ("not running" in stderr_lower or "stopped" in stderr_lower),
+        ]
+        return any(service_not_running_patterns)
+
     def pull_image(self, image: Optional[str] = None, force_pull: bool = False) -> bool:
-        """Pull a container image if not already available or force_pull is True."""
-        pass
+        """Pull a container image if not already available or force_pull is True.
+
+        Args:
+            image: Image name to pull (defaults to default_image)
+            force_pull: Force pull even if image exists locally
+
+        Returns:
+            True if image is available after operation, False otherwise
+
+        Raises:
+            ContainerImagePullError: If image pull fails with details
+            ContainerTimeoutError: If pull operation times out
+            ContainerPermissionError: If permission denied during pull
+        """
+        import logging
+        import subprocess
+
+        logger = logging.getLogger(__name__)
+        target_image = image or self.default_image
+
+        # Import exceptions here to avoid circular imports
+        from .exceptions import (
+            ContainerImagePullError,
+            ContainerPermissionError,
+            ContainerTimeoutError,
+        )
+
+        # If force_pull is False, check if image is already available locally
+        if not force_pull:
+            try:
+                result = subprocess.run(
+                    [self.engine_name, "image", "inspect", target_image],
+                    capture_output=True,
+                    text=True,
+                    timeout=10,
+                )
+                if result.returncode == 0:
+                    logger.debug(f"{self.engine_name.title()} image {target_image} already available locally")
+                    return True  # Image already available locally
+            except subprocess.TimeoutExpired:
+                logger.debug(f"Timeout checking local image {target_image}, proceeding with pull")
+            except subprocess.CalledProcessError:
+                logger.debug(f"Image {target_image} not available locally, proceeding with pull")
+
+        # Pull the latest version of the image
+        logger.info(f"Pulling {self.engine_name.title()} image: {target_image}")
+        try:
+            result = subprocess.run(
+                [self.engine_name, "pull", target_image],
+                capture_output=True,
+                text=True,
+                timeout=300,  # 5 minutes
+            )
+
+            if result.returncode == 0:
+                logger.info(f"Successfully pulled {self.engine_name.title()} image: {target_image}")
+                return True
+            else:
+                # Analyze the error to provide helpful feedback
+                stderr_lower = result.stderr.lower()
+                if "permission denied" in stderr_lower:
+                    raise ContainerPermissionError(self.engine_name, f"pull image {target_image}")
+                elif "not found" in stderr_lower or "no such image" in stderr_lower:
+                    raise ContainerImagePullError(self.engine_name, target_image, "Image not found in registry")
+                elif "network" in stderr_lower or "connection" in stderr_lower:
+                    raise ContainerImagePullError(self.engine_name, target_image, "Network connectivity issue")
+                elif "unauthorized" in stderr_lower or "authentication" in stderr_lower:
+                    raise ContainerImagePullError(
+                        self.engine_name, target_image, "Authentication required for private image"
+                    )
+                else:
+                    raise ContainerImagePullError(self.engine_name, target_image, result.stderr.strip())
+
+        except subprocess.TimeoutExpired as e:
+            raise ContainerTimeoutError(self.engine_name, f"pull image {target_image}", 300) from e
+        except subprocess.CalledProcessError as e:
+            logger.debug(f"{self.engine_name.title()} pull failed with exit code {e.returncode}")
+            raise ContainerImagePullError(
+                self.engine_name, target_image, f"Pull command failed with exit code {e.returncode}"
+            ) from e
 
     @abstractmethod
     def run_command(
@@ -573,3 +730,33 @@ except ImportError as e:
 
         except (subprocess.TimeoutExpired, subprocess.CalledProcessError):
             return False
+
+    def _cleanup_expired_sessions(self, force: bool = False) -> None:
+        """Clean up expired or inactive container sessions."""
+        import time
+
+        current_time = time.time()
+
+        # Only run cleanup every 30 seconds unless forced
+        if not force and current_time - self._last_cleanup < 30:
+            return
+
+        self._last_cleanup = current_time
+        expired_keys = []
+
+        for key, session in self._active_sessions.items():
+            session_age = current_time - (session.created_at or 0.0)
+            if session_age > self._session_timeout or not session.is_active():
+                session.cleanup()
+                expired_keys.append(key)
+
+        for key in expired_keys:
+            del self._active_sessions[key]
+
+        # If we have too many sessions, cleanup the oldest ones
+        if len(self._active_sessions) > self._max_sessions:
+            sorted_sessions = sorted(self._active_sessions.items(), key=lambda x: x[1].created_at or 0.0)
+            excess_count = len(self._active_sessions) - self._max_sessions
+            for key, session in sorted_sessions[:excess_count]:
+                session.cleanup()
+                del self._active_sessions[key]
